Always set provided EXI configuration options
Whenever a client sends a EXI configration property, use it explicitly.

Previous code assumed that 'false' was the default state of properties. Even if true, this might change in the future. Issues are avoided by explicitly setting whatever value is provided by the client, rather than implicitly depending on a default value.